+++ src/basic/features/index.js
@@ -6,4 +6,5 @@ export * from './events/index.js';
 export { setupObservers } from './observerFactory.js';
 export * from './points/index.js';
 export * from './product/index.js';
+export * from './sales/index.js';
 export * from './ui/index.js';
